% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_sets}
\alias{gene_sets}
\title{Gene sets to infer tumor stromal and immune infiltration}
\format{
A data frame with 141 row and 2 variables:
  \describe{
    \item{stromal_signature}{Geneset of HGNC symbols used to infer tumor stromal cell infiltration}
    \item{immune_signature}{Geneset of HGNC symbols used to infer tumor immune cell infiltration}
  }
}
\source{
\url{https://r-forge.r-project.org/scm/viewvc.php/pkg/estimate/data/SI_geneset.RData?root=estimate&view=log}
}
\usage{
gene_sets
}
\description{
Two gene sets, each 141 genes in length, created to infer
  stromal and immune infiltration
}
\keyword{datasets}
