% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ov}
\alias{ov}
\title{Ovarian cancer tumor RNA expression}
\format{
A matrix with 17256 rows and 10 columns, where each column represents
  a tumor, and each row represents a gene. Genes are represented by HGNC
  symbols.
}
\source{
\url{https://r-forge.r-project.org/scm/viewvc.php/pkg/estimate/inst/extdata/sample_input.txt?root=estimate&view=log}
}
\usage{
ov
}
\description{
A matrix containing RNA expression of 10 ovarian cancer tumors,
  measured using the Affymetrix U133Plus2.0 platform. These data have been
  rounded to the 4th decimal place to reduce file size.
}
\keyword{datasets}
