% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{purity_data_affy}
\alias{purity_data_affy}
\title{Affymetrix data used to train ESTIMATE algorithm}
\format{

}
\usage{
purity_data_affy
}
\description{
A dataframe containing the ABSOLUTE-measured and
  ESTIMATE-predicted purity values of 995 tumors. Additionally, stromal and
  immune scores as calculated by ESTIMATE. All tumors were profiled on
  Affymetrix arrays, and were used to generate the Affymetrix algorithm.
}
\keyword{datasets}
