% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyestimate.R
\docType{package}
\name{tidyestimate}
\alias{tidyestimate}
\title{tidyestimate: A modern implementation of the ESTIMATE algorithm}
\description{
The tidyestimate is a lightweight, fast, pipe-friendly re-imagination of the
ESTIMATE package. tidyestimate is used to infer tumor purity from
expression data.
}
\section{Authors}{

Author (tidyestimate):

* Kai Aragaki ([ORCID](http://orcid.org/0000-0002-9458-0426)) (author, maintainer)

Authors (ESTIMATE):

* Kosuke Yoshihara kyoshihara@mdanderson.org (author)
* P. Roebuck proebuck@mdanderson.org (author, copyright holder)
}

\section{Reference}{

https://www.nature.com/articles/ncomms3612
}

