% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_score.R
\name{estimate_score}
\alias{estimate_score}
\title{Calculate Stromal, Immune, and ESTIMATE scores}
\usage{
estimate_score(df, is_affymetrix, tidy)
}
\arguments{
\item{df}{a dataframe of normalized expression data, where columns are
patients and genes are rows. Gene names must be HGNC, and may either be the
first column or the rownames (see `tidy` argument below)}

\item{is_affymetrix}{logical. Is the expression data from an Affymetrix
array?}

\item{tidy}{logical. If true, assumes the first column contains HGNC symbols.
If false, assumes rownames contain HGNC symbols}
}
\value{
A data frame with scores for stromal, immune, and ESTIMATE scores per tumor.
}
\description{
Calculate Stromal, Immune, and ESTIMATE scores
}
